Fix stderr not being checked in test1
The default `failed` string is present in stderr and not stdout, hence it is not being sent to the autograder currently. This PR fixes the issue by redirecting stderr to stdout and then exiting with error code 0 (the autograder rejects solutions exiting with non-zero error code by default and failing in picat emits error code 1).